Levenshtein: Generalize tracing a recursive path
Make _reconstructPath public and generalize it, letting the user pass a function
in so that they can do what they want with the data on each step through the
path. This is useful for e.g. getting a count of each type of operation: insert;
delete; match; substitute, without having to loop through the reconstructed
path itself. It can be done on the first pass, and only the work the
user wants done is done.