Levenshtein: insertCost now passed boundary chars
The insertCost function now is provided two extra arguments: the
character in the search string to the left of the cursor position of the
insertion, and likewise the character to the right of that position.
This allows the user to set the cost based on the position of the
insertion inside the search pattern, such as within or outside the
boundary of a word.